requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
jwt==1.3.1
pyjwt==2.8.0
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1